cmd/logger: isolate logger code into its own file
